added pk class
